% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker.R
\name{add_assets}
\alias{add_assets}
\title{Add assets for the specified use case}
\usage{
add_assets(
  pkg_path = ".",
  img_path = NULL,
  use_case = "default",
  overwrite = TRUE
)
}
\arguments{
\item{pkg_path}{Path to the package directory}

\item{img_path}{Path to the write the docker image definition contents; default \code{NULL} will use \verb{docker/} as a sub-directory of the "pkg_path"}

\item{use_case}{Name of the use case. Defaults to \code{"default"}, which only uses the base boilerplate.}

\item{overwrite}{Logical as to whether or not existing assets should be overwitten; default is \code{TRUE}}
}
\value{
Invisibly returns assets per \link{handle_use_case}. Called primarily for its side effects.
}
\description{
Add assets for the specified use case
}
\examples{
\dontrun{
add_assets(use_case="shiny")
add_assets(use_case="pipeline")
}
}
