% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker.R
\name{add_dockerfile}
\alias{add_dockerfile}
\title{Add a Dockerfile to the docker directory}
\usage{
add_dockerfile(
  path = ".",
  base_image = "rocker/r-ver:latest",
  pkgs = NULL,
  use_renv = TRUE
)
}
\arguments{
\item{path}{Path to the package directory}

\item{base_image}{Name of base image to start FROM in Dockerfile}

\item{pkgs}{Vector of packages to include in Dockerfile; only relevant if \code{use_renv = FALSE}}

\item{use_renv}{Logical as to whether or not to use renv}
}
\value{
A list with information about the package. Also called for side-effect, creates Dockerfile.
}
\description{
Add a Dockerfile to the docker directory
}
\examples{
\dontrun{
add_dockerfile()
}
}
