% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{build_pkg}
\alias{build_pkg}
\title{Build a package tar.gz}
\usage{
build_pkg(dir = "docker")
}
\arguments{
\item{dir}{Location for the built package tar.gz. Defaults to \verb{docker/}}
}
\value{
Nothing
}
\description{
Build a package tar.gz and move it into a user-specified location (default \verb{docker/})
}
\examples{
\dontrun{
build_pkg()
}
}
