% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker.R
\name{create_ddir}
\alias{create_ddir}
\title{Function to create docker directory}
\usage{
create_ddir(path = getwd())
}
\arguments{
\item{path}{Path to the package directory}
}
\value{
Side-effect. Creates directory.
}
\description{
Function to create docker directory
}
\examples{
\dontrun{
create_ddir()
}

}
