% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{handle_use_case}
\alias{handle_use_case}
\title{Handle the use case}
\usage{
handle_use_case(use_case)
}
\arguments{
\item{use_case}{The specified use case.}
}
\value{
List of parsed information for the use case including, the name of the use case, path to Dockerfile template, base image, and path to assets (delimited by \verb{;} if there are multiple and \code{NA} if there are none).
}
\description{
This unexported helper function internally handles the provided use case.
}
