% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pkg_root}
\alias{pkg_root}
\title{Find package root}
\usage{
pkg_root(pkg_path = ".", ...)
}
\arguments{
\item{pkg_path}{Path to the package directory.}

\item{...}{Arguments passed to \link[rprojroot:find_root_file]{rprojroot::find_package_root_file}.}
}
\value{
A file path of the package root.
}
\description{
Find package root
}
