% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pkginfo}
\alias{pkginfo}
\title{Get information about the current package}
\usage{
pkginfo(...)
}
\arguments{
\item{...}{Arguments passed to \link[rprojroot:find_root_file]{rprojroot::find_package_root_file}. E.g. \code{path="/path/to/dir"}.}
}
\value{
A list fixme
}
\description{
Returns information about the current package in a list which can be passed to other functions.
}
\examples{
\dontrun{
pkginfo()
}
}
