% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pracpac-package.R
\docType{package}
\name{pracpac-package}
\alias{pracpac}
\alias{pracpac-package}
\title{pracpac: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: VP Nagraj \email{pnagraj@signaturescience.com} (\href{https://orcid.org/0000-0003-0060-566X}{ORCID})

Authors:
\itemize{
  \item Stephen Turner (\href{https://orcid.org/0000-0001-9140-9028}{ORCID})
}

Other contributors:
\itemize{
  \item Signature Science, LLC. [copyright holder]
}

}
\keyword{internal}
