% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker.R
\name{use_docker}
\alias{use_docker}
\title{Title}
\usage{
use_docker(
  path = ".",
  use_renv = TRUE,
  base_image = "rocker/r-ver:latest",
  other_packages = NULL,
  build = TRUE
)
}
\arguments{
\item{build}{}
}
\description{
Title
}
