% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker.R
\name{add_dockerfile}
\alias{add_dockerfile}
\title{Title}
\usage{
add_dockerfile(
  path = getwd(),
  base_image = "rocker/r-ver:latest",
  pkgs = NULL,
  use_renv = TRUE
)
}
\arguments{
\item{use_renv}{}
}
\description{
Title
}
