% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker.R
\name{add_dockerfile}
\alias{add_dockerfile}
\title{Add a Dockerfile to the docker directory}
\usage{
add_dockerfile(
  path = ".",
  base_image = "rocker/r-ver:latest",
  use_renv = TRUE,
  usecase = NULL,
  repos = NULL
)
}
\arguments{
\item{path}{Path to the package directory}

\item{base_image}{Name of base image to start \code{FROM} in Dockerfile}

\item{use_renv}{Logical as to whether or not to use renv. Defaults to \code{TRUE}. If \code{FALSE}, package dependencies are scraped from the \code{DESCRIPTION} file and the most recent versions will be installed in the image.}

\item{usecase}{One of the use case templates in inst/templates. Defaults to \code{NULL} -- no additional Dockerfile boilerplate is added.}

\item{repos}{Option to override the repos used for installing packages with \code{renv} by passing name of repository. Only used if \code{use_renv = TRUE}. Default is \code{NULL} meaning that the repos specified in \code{renv} lockfile will remain as-is and not be overridden.}
}
\value{
A list with information about the package. Also called for side-effect, creates Dockerfile.
}
\description{
Add a Dockerfile to the docker directory
}
\examples{
\dontrun{
add_dockerfile(base_image="rocker/r-ver:4.2.2", use_renv=TRUE)
add_dockerfile(base_image="rocker/r-ver:4.2.2", use_renv=TRUE, usecase="helloworld")
add_dockerfile(base_image="rocker/r-ver:4.2.2", use_renv=FALSE)
add_dockerfile(base_image="rocker/r-ver:4.2.2", use_renv=FALSE, usecase="helloworld")
}
}
