% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{build_image}
\alias{build_image}
\title{Build a Docker image}
\usage{
build_image(pkg_path = ".", img_path = NULL, cache = TRUE)
}
\arguments{
\item{pkg_path}{Path to the package directory}

\item{img_path}{Path to the write the docker image definition contents; default \code{NULL} will use \verb{docker/} as a sub-directory of the "pkg_path"}

\item{cache}{Logical; should caching be used? Default \code{TRUE}. Set to \code{FALSE} to use \code{--no-cache} in \verb{docker build}.}
}
\value{
(Invisible) The \verb{docker build} command. Called for its side effects, which runs the \verb{docker build} as a system command.
}
\description{
Build a Docker image created by FIXME function name.
}
\examples{
\dontrun{
build_image()
}
}
