% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{build_image}
\alias{build_image}
\title{Build a Docker image}
\usage{
build_image(path = ".", build = TRUE, cache = TRUE)
}
\arguments{
\item{path}{Path to a package directory containing a \code{docker} subdirectory, which contains the Dockerfile built by \link{add_dockerfile}}

\item{build}{Logical; should the image actually be built? Default \code{TRUE}. Set to \code{FALSE} for debugging.}

\item{cache}{Logical; should caching be used? Default \code{TRUE}. Set to \code{FALSE} to use \code{--no-cache} in \verb{docker build}.}
}
\value{
Nothing
}
\description{
Build a Docker image created by FIXME function name.
}
\examples{
\dontrun{
build_image(build=FALSE)
build_image()
}
}
