% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{build_pkg}
\alias{build_pkg}
\title{Build a package tar.gz}
\usage{
build_pkg(docker_dir = "docker", build = TRUE)
}
\arguments{
\item{docker_dir}{Location for the built package tar.gz. Defaults to \verb{docker/}}

\item{build}{Logical; should the package actually be built? Default \code{TRUE}. Set to \code{FALSE} for debugging.}
}
\value{
A list of package info returned by \link{pkginfo}, tar.gz source and destination file paths.
}
\description{
Build a package tar.gz and move it into a user-specified location (default \verb{docker/})
}
\examples{
\dontrun{
build_pkg(docker_dir="docker", build=FALSE)
build_pkg(docker_dir="docker", build=TRUE)
}
}
