% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{build_pkg}
\alias{build_pkg}
\title{Build a package tar.gz}
\usage{
build_pkg(pkg_path = ".", img_path = NULL, ...)
}
\arguments{
\item{pkg_path}{Path to the package directory}

\item{img_path}{Path to the write the docker image definition contents; default \code{NULL} will use \verb{docker/} as a sub-directory of the "pkg_path"}

\item{...}{Additional optional arguments passed to \link[pkgbuild:build]{pkgbuild::build}}
}
\value{
(Invisible) A list of package info returned by \link{pkg_info}, tar.gz source and destination file paths.
}
\description{
Build a package tar.gz and move it into a user-specified location (default \verb{docker/})
}
\examples{
\dontrun{
build_pkg()
}
}
