% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker.R
\name{create_docker_dir}
\alias{create_docker_dir}
\title{Function to create docker directory}
\usage{
create_docker_dir(pkg_path = ".", img_path = NULL)
}
\arguments{
\item{pkg_path}{Path to the package directory}

\item{img_path}{Path to the write the docker image definition contents; default \code{NULL} will use \verb{docker/} as a sub-directory of the "pkg_path"}
}
\value{
(Invisible) A list of package info returned by \link{pkg_info}. Also called for side-effect, creates docker directory.
}
\description{
Function to create docker directory
}
\examples{
\dontrun{
create_docker_dir()
}

}
