% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker.R
\name{create_docker_dir}
\alias{create_docker_dir}
\title{Function to create docker directory}
\usage{
create_docker_dir(path = ".")
}
\arguments{
\item{path}{Path to the package directory}
}
\value{
(Invisible) A list with information about the package. Also called for side-effect, creates docker directory.
}
\description{
Function to create docker directory
}
\examples{
\dontrun{
create_docker_dir()
}

}
