% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pkginfo}
\alias{pkginfo}
\title{Get information about the current package}
\usage{
pkginfo(path = ".", ...)
}
\arguments{
\item{path}{Directory path to a package root}

\item{...}{Arguments passed to \link[rprojroot:find_root_file]{rprojroot::find_package_root_file}}
}
\value{
A list fixme
}
\description{
Returns information about the current package in a list which can be passed to other functions.
}
\examples{
\dontrun{
# This will succeed if this is a package
pkginfo()
# This will fail if this is not a package location
pkginfo(tempdir())
}
}
