% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pracpac-package.R
\docType{package}
\name{pracpac-package}
\alias{pracpac}
\alias{pracpac-package}
\title{pracpac: Practical 'R' Packaging in 'Docker'}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Streamline the creation of 'Docker' images with 'R' packages and dependencies embedded. The 'pracpac' package provides a 'usethis'-like interface to creating Dockerfiles with dependencies managed by 'renv'. The 'pracpac' functionality is described in Nagraj and Turner (2023) \doi{10.48550/arXiv.2303.07876}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://signaturescience.github.io/pracpac/}
  \item \url{https://github.com/signaturescience/pracpac/}
  \item Report bugs at \url{https://github.com/signaturescience/pracpac/issues}
}

}
\author{
\strong{Maintainer}: VP Nagraj \email{nagraj@nagraj.net} (\href{https://orcid.org/0000-0003-0060-566X}{ORCID})

Authors:
\itemize{
  \item Stephen Turner (\href{https://orcid.org/0000-0001-9140-9028}{ORCID})
}

Other contributors:
\itemize{
  \item Signature Science, LLC. [copyright holder]
}

}
\keyword{internal}
