% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker.R
\name{renv_deps}
\alias{renv_deps}
\title{Get renv dependencies}
\usage{
renv_deps(path = ".", other_packages = NULL)
}
\arguments{
\item{path}{Path to the package directory}

\item{other_packages}{Vector of other packages to be included in \code{renv} lock file; default is \code{NULL}}
}
\value{
(Invisible) A list with information about the package. Primarily called for side effect. Writes an \code{renv} lock file to the docker/ directory.
}
\description{
Get renv dependencies
}
\examples{

\dontrun{
renv_deps()
}
}
